Add fibertree-bootstrap submodule and update all submodules
The fibertree-bootstrap submodule is a Python package that can
universally bootstrap a fibertree notebook environment either running
in Juypyter or Google Colab. This commit adds it as a submodule,
installs it as part of the Docker image along with its documentation.

Also updated all the submodules to their respective HEADs.